# Add #[inline(always)] to improve compile performance of v1