Benchmarks show a performance improvement by roughly 5%.
More performance improvements are expected to be gained by transitioning to the wasmparser crate.