@Robbepop 2022-01-09
> I trust you, nonetheless it would be interesting to see the benchmark outputs.

Good point!
Brace yourself for ultimate confusion though ...
I just found out that `wasmi` benchmark with respect to `compile_and_validate` behaves very differently if executed together with other benchmarks or not, i.e. when doing `cargo bench` versus `cargo bench compile_and_validate`. :facepalm: 

We really do need Criterion support ...

Well in both cases we can see the roughly 5% performance improvement at least.
I will post all combinations below.

# Benchmarks

## Before

### Isolation

```
test bench_compile_and_validate    ... bench:   7,263,582 ns/iter (+/- 402,986)
test bench_compile_and_validate_v1 ... bench:   7,360,928 ns/iter (+/- 345,759)
```

### Combined

```
test bench_compile_and_validate    ... bench:   7,179,156 ns/iter (+/- 299,672)
test bench_compile_and_validate_v1 ... bench:   8,351,575 ns/iter (+/- 439,091)
```

## After

### Isolation

```
test bench_compile_and_validate    ... bench:   7,119,371 ns/iter (+/- 500,738)
test bench_compile_and_validate_v1 ... bench:   6,965,588 ns/iter (+/- 303,476)
```

### Combined

```
test bench_compile_and_validate    ... bench:   7,143,611 ns/iter (+/- 432,086)
test bench_compile_and_validate_v1 ... bench:   7,969,667 ns/iter (+/- 430,899)
```